Add `disable_inheritance_instantiation` setting
When loading a lot of objects from the top of the inheritance chain, to
load attributes of the subclass instances an additional DB lookup query
is made. To prevent this, `self.disable_inheritance_instantiation = true`
can be set in the topmost class and the number of DB lookups will stay
the same, however, instances will be missing attributes.